base-requirements.txt
poetry
mypy
black
isort
pytest
pytest-asyncio
mypy_extensions
typing_extensions
python-dotenv
